Refactor workflow tracing: move tracing logic to a new module and update imports
