+++ src/workflow.py
@@ -6,10 +6,9 @@ from llama_index.core.workflow import (
     Context,
     step,
 )
-from langfuse.llama_index import LlamaIndexInstrumentor
-
 
 from src.config import config
+from src.workflow import AssistantFlow
 
 
 class PreprocessEvent(Event):
@@ -26,14 +25,6 @@ class DeduplicateEvent(Event):
 
 class SanityCheckEvent(Event):
     qa: list[tuple[str, str]]
-
-
-# Initialize the Langfuse instrumentor
-instrumentor = LlamaIndexInstrumentor(
-    public_key=config["public_key"],
-    secret_key=config["secret_key"],
-    host=config["host"],
-)
 
 
 class AssistantFlow(Workflow):
@@ -96,27 +87,3 @@ class AssistantFlow(Workflow):
         result = await reply(query_clean, qa)
 
         return StopEvent(result=result)
-
-
-# Example of how to use the workflow with Langfuse tracing
-async def run_workflow_with_tracing(
-    query: str, session_id: str = None, user_id: str = None
-):
-    # Start the instrumentation
-    instrumentor.start()
-
-    # Or use the context manager for more control over tracing parameters
-    with instrumentor.observe(
-        trace_id=f"assistant-flow-{query[:10]}",  # Optional custom trace ID
-        session_id=session_id,
-        user_id=user_id,
-        metadata={"original_query": query},
-    ) as trace:
-        # Run your workflow
-        workflow = AssistantFlow(timeout=3 * 60)
-        result = await workflow.run(query=query)
-
-    # Make sure to flush before the application exits
-    instrumentor.flush()
-
-    return result
